requirements.txt
matplotlib
numpy
pyvisa